iot_simulator/fog_nodes/code/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
pandas
pandas-gbq
db-dtypes
